requirements.txt
pyserial
numpy
opencv-python
requests
python-dotenv